Split long running test up so it can be paralllelised.  Enable reference tests on PRs to see how long they now take.
