+++ .circleci/config.yml
@@ -166,11 +166,6 @@ workflows:
       - referenceTests:
           requires:
             - assemble
-          filters:
-            branches:
-              only:
-                - master
-                - /^release-.*/
       - unitTests:
           requires:
             - assemble
